Rename validation schema file for measurements back to measurements to reflect service name
