+++ backend/src/validation/measurement.ts
